+++ log/log.go
@@ -1,0 +1,2 @@
+// Package log extends the log package in the stdlib.
+package log